Finished project 3 (second part of KENSv3)
This is the version that uses custom constructors and dynamic memory allocation, which was done because of a mistaken analysis of a problem.
An attempt will be made to create a simplified version without the custom constructors and dynamic memory allocation.
If it works, then it will be pushed to this repository as the final form of project 3.